+++ libyul/backends/evm/EthAssemblyAdapter.cpp
@@ -232,12 +232,12 @@ void EthAssemblyAdapter::appendAuxDataLoadN(uint16_t _offset)
 	m_assembly.appendAuxDataLoadN(_offset);
 }
 
-void EthAssemblyAdapter::appendSwapN(uint8_t _depth)
+void EthAssemblyAdapter::appendSwapN(size_t _depth)
 {
 	m_assembly.appendSwapN(_depth);
 }
 
-void EthAssemblyAdapter::appendDupN(uint8_t _depth)
+void EthAssemblyAdapter::appendDupN(size_t _depth)
 {
 	m_assembly.appendDupN(_depth);
 }
